🎨 Improve Counter Structure
