+++ webpack.config.js
@@ -1,7 +1,7 @@
 const path = require('path');
 
 module.exports = {
-  entry: path.resolve(__dirname, 'src/page.jsx'),
+  entry: path.resolve(__dirname, 'src/index.jsx'),
   module: {
     rules: [
       {
